renaming migrations csproj
