+++ DotNetRuServer.Migrations/11_RemoveNotNullFriendLogo.cs
@@ -1,6 +1,6 @@
 using FluentMigrator;
 
-namespace DevActivator.Migrations
+namespace DotNetRuServer.Migrations
 {
     [Migration(11)]
     public class RemoveNotNullFriendLogo : Migration
