+++ DotNetRuServer.Migrations/3_AddSpeakerTalkTable.cs
@@ -1,6 +1,6 @@
 using FluentMigrator;
 
-namespace DevActivator.Migrations
+namespace DotNetRuServer.Migrations
 {
     [Migration(3)]
     public class AddSpeakerTalkTable : Migration
