+++ ElectronNetAngular.sln
@@ -17,7 +17,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DevActivator", "DevActivato
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Database", "Database", "{914AA934-D2CE-4E5E-87E9-AF9F8F6B9B06}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DevActivator.Migrations", "DevActivator.Migrations\DevActivator.Migrations.csproj", "{0FF6C9FB-BC4E-4D5A-AEF1-4E6ADE93990B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotNetRuServer.Migrations", "DotNetRuServer.Migrations\DotNetRuServer.Migrations.csproj", "{0FF6C9FB-BC4E-4D5A-AEF1-4E6ADE93990B}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Importer", "Importer\Importer.csproj", "{88572F7C-12FD-4374-A612-A8962EC98F39}"
 EndProject
